+++ syscall.c
@@ -48,6 +48,7 @@
 #endif
 
 #include "ifuncs.h"
+#include "vfs/vfs_internal.h"
 
 extern int dry_run;
 extern int am_root;
@@ -66,29 +67,7 @@ extern int am_daemon;
 extern int am_chrooted;
 extern int insecure_links;
 extern int module_id;
-extern unsigned int module_dirlen;
-extern char *module_dir;
-extern int module_dirfd;	/* daemon: served module root pinned by identity, or -1 */
-extern char *confine_root;	/* --confine-root, or NULL; see confinement_root() */
-extern unsigned int confine_rootlen;
-extern char curr_dir[MAXPATHLEN];	/* defined below; fwd-declared for the seed */
-extern int operator_path_resolve;	/* defined below; fwd-declared for the exclude check */
 
-#if defined AT_FDCWD && defined O_NOFOLLOW && defined O_DIRECTORY
-/* Open a trusted absolute anchor directory as an owned dirfd.  When the anchor is
- * the served module root and the daemon pinned it by identity (module_dirfd), dup
- * that fd rather than re-resolving the absolute path with openat(AT_FDCWD, ...) --
- * which re-traverses the module's ancestors as the dropped-privilege module uid
- * and EACCESes when the module sits under a non-traversable parent (a 0700 home).
- * Functionally identical (same inode), just privilege-drop-safe.  Gated like its
- * callers (the secure resolver and dpc_dir_fd both require these three). */
-static int open_anchor_dirfd(const char *path)
-{
-	if (module_dirfd >= 0 && am_daemon && module_dir && strcmp(path, module_dir) == 0)
-		return dup(module_dirfd);
-	return openat(AT_FDCWD, path, O_RDONLY | O_DIRECTORY);
-}
-#endif
 
 /* Single gate for whether path resolution must be hardened against
  * parent-component symlink races (TOCTOU).  Used by the do_*_at()/do_*_atfd()
@@ -126,118 +105,6 @@ int symlink_optout_allowed(void)
 	return insecure_links;
 }
 
-/* The root an operator/peer-supplied path must stay under, or NULL when nothing
- * is confined.  A daemon has the served module; a server launched by a wrapper
- * with its own restricted directory (rrsync) gets one from --confine-root.
- *
- * A daemon never honours --confine-root: module_dir is the boundary there, and
- * the option arrives in a peer-supplied argv, so obeying it could only loosen
- * the module. */
-static const char *confinement_root(unsigned int *lenp)
-{
-	if (am_daemon) {
-		*lenp = module_dirlen;
-		return module_dir;
-	}
-	*lenp = confine_rootlen;
-	return confine_root;
-}
-
-/* Split the "/proc/<self|pid>/fd" prefix off `p`, returning the tail -- "" for
- * the pin directory itself, otherwise a string starting with '/'.  NULL when `p`
- * is not in the fd-pin namespace at all. */
-static const char *fd_pin_tail(const char *p)
-{
-	const char *s;
-
-	if (strncmp(p, "/proc/", 6) != 0)
-		return NULL;
-	s = p + 6;
-	if (strncmp(s, "self/", 5) == 0)	/* "/proc/self/..." */
-		s += 4;
-	else {					/* "/proc/<pid>/..." */
-		const char *d = s;
-		while (*s >= '0' && *s <= '9')
-			s++;
-		if (s == d || *s != '/')
-			return NULL;
-	}
-	if (strncmp(s, "/fd", 3) != 0)
-		return NULL;
-	s += 3;
-	return (*s == '\0' || *s == '/') ? s : NULL;
-}
-
-/* An EXACT pin entry, "/proc/self/fd/7" -- the one spelling whose target is what
- * confinement must judge.  rrsync also writes a pinned parent as
- * ".../fd/7/<leaf>", but the walk resolves the magic link itself and checks the
- * components past it, so only the bare entry is resolved here.  Requiring all
- * digits keeps a planted name like ".../fd/outside-secret" out. */
-static int is_exact_fd_pin(const char *p)
-{
-	const char *tail = fd_pin_tail(p);
-
-	if (!tail || *tail != '/')
-		return 0;
-	for (++tail; *tail >= '0' && *tail <= '9'; tail++) {}
-	return *tail == '\0' && tail[-1] != '/';
-}
-
-/* Refuse (return 1) when the ABSOLUTE resolved path `abspath` lands OUTSIDE the
- * confinement root, for an operator/peer-supplied path that must stay inside it
- * (--partial-dir/--backup-dir/alt-basis/merge files: operator_path_resolve).  An
- * in-tree symlink owned by uid 0 / the euid is followed by design, so it can
- * redirect the resolved target outside the root; this catches that escape.
- *
- * This is ROOT confinement only.  The daemon exclude/filter list is a name-based
- * visibility filter, NOT a physical-path boundary: a symlink whose own name is
- * not excluded may still resolve into an excluded IN-tree subtree, exactly as in
- * stock rsync.  The defense for a writable module is `munge symlinks` (see
- * rsyncd.conf(5)), not this walk. */
-static int abspath_outside_confinement(const char *abspath)
-{
-	unsigned int rootlen;
-	const char *root = confinement_root(&rootlen);
-	char pinned[MAXPATHLEN];
-
-	if (!root || !abspath)
-		return 0;
-	if (rootlen <= 1)			/* root is "/": nothing is outside */
-		return 0;
-	/* An fd pin (rrsync rewrites a validated option path to /proc/self/fd/N so
-	 * no later symlink can redirect it) is spelled outside the root by
-	 * construction.  Judge it by what it points AT rather than by its spelling,
-	 * so a pin is neither wrongly refused nor blindly trusted.  A pin we cannot
-	 * resolve to an absolute path is refused, not waved through: an unreadable
-	 * pin is exactly the case where we cannot say where the open would land. */
-	if (!am_daemon) {
-		const char *tail = fd_pin_tail(abspath);
-		if (tail && !*tail)
-			return 0;		/* the pin directory: transit, opens nothing */
-		if (is_exact_fd_pin(abspath)) {
-			ssize_t n = readlink(abspath, pinned, sizeof pinned - 1);
-			if (n <= 0 || pinned[0] != '/')
-				return operator_path_resolve ? 1 : 0;
-			pinned[n] = '\0';
-			abspath = pinned;
-		}
-	}
-	if (strncmp(abspath, root, rootlen) == 0
-	 && (abspath[rootlen] == '\0' || abspath[rootlen] == '/'))
-		return 0;			/* inside: name-based exclude is not a boundary */
-	/* Not under the root.  An ABSOLUTE walk passes through the root's ancestors
-	 * ("/", "/home", ...) on the way down -- those are not "outside", just
-	 * not-yet-arrived, so allow them.  A path that has truly DIVERGED is
-	 * outside: refuse it for an operator/peer path that must stay in the tree
-	 * (operator_path_resolve); other opens (--log-file, --*-from, lock/motd)
-	 * may legitimately live elsewhere.  The --insecure-links / "insecure links
-	 * = yes" opt-out short-circuits before we get here. */
-	size_t alen = strlen(abspath);
-	if (alen == 0
-	 || (strncmp(abspath, root, alen) == 0 && root[alen] == '/'))
-		return 0;			/* ancestor of the root: still descending */
-	return operator_path_resolve ? 1 : 0;
-}
 
 /* Advance the tracked absolute path `abspath` by one resolved component,
  * normalizing "." and ".." exactly as openat() does so the module-confinement
@@ -2747,27 +2614,6 @@ int do_open_nofollow(const char *pathname, int flags)
   (except for a deliberately re-anchored module path; see below).
 */
 
-/* Returns 1 if path has any "/"-separated component that is exactly
- * "..", 0 otherwise. Used by secure_relative_open's front-door
- * validation to reject ".." inputs (bare "..", "foo/..", "subdir/..")
- * for non-re-anchored paths; the walk itself resolves an in-tree ".."
- * safely (ds_descend pops to the held parent) for a re-anchored path. */
-static int path_has_dotdot_component(const char *path)
-{
-	const char *p = path;
-
-	while (*p) {
-		const char *q;
-		if (*p == '/') { p++; continue; }
-		q = p;
-		while (*q && *q != '/')
-			q++;
-		if (q - p == 2 && p[0] == '.' && p[1] == '.')
-			return 1;
-		p = q;
-	}
-	return 0;
-}
 
 /* The logical current directory (maintained by change_dir() in util1.c).
  * Defined here -- rather than in util1.c -- so the test helpers that link
@@ -2777,201 +2623,6 @@ char curr_dir[MAXPATHLEN];
 unsigned int curr_dir_len;
 
 #if defined(O_NOFOLLOW) && defined(O_DIRECTORY) && defined(AT_FDCWD)
-/* In-tree symlink following for secure_relative_open()'s directory walk (below).
- * An early version refused every symlink with O_NOFOLLOW on each component, which
- * broke legitimate within-tree directory symlinks (--keep-dirlinks #715, and -aR
- * through a symlinked parent); the walk now follows them safely.
- *
- * A directory walk keeps a stack of the open dirfds from the anchor (index 0,
- * borrowed -- not closed here) down to the current directory.  Descending into
- * a real subdirectory pushes its fd; a ".." in a followed symlink target pops
- * back to the already-pinned parent fd rather than re-resolving ".." with
- * openat(), so an ancestor renamed mid-walk cannot redirect the climb, and the
- * climb can never rise above the anchor (a pop at the anchor returns ELOOP).
- * This matches RESOLVE_BENEATH, which allows in-tree ".." that stays beneath the
- * root.  Absolute symlink targets are refused; symlink hops are bounded. */
-#ifndef SECURE_OPEN_MAXSYMLINKS
-#define SECURE_OPEN_MAXSYMLINKS 40
-#endif
-
-/* Max directory levels held open at once during a single resolve.  The walk
- * holds one fd per component, so depth is bounded by RLIMIT_NOFILE anyway; a
- * fixed array (no malloc/realloc) keeps the stack simple and the static
- * analyzer happy.  Mirrors DPC_MAXDEPTH's fixed-cap approach. */
-#define DS_MAXDEPTH 1024
-
-struct dirstack {
-	int fds[DS_MAXDEPTH];	/* fds[0] = anchor (borrowed); fds[top] = current dir */
-	int top;
-	/* Absolute path of fds[top], maintained as we descend/pop, for the
-	 * exclude-aware refusal (abspath_outside_confinement).  Empty unless the
-	 * caller seeds it with the anchor's absolute path; then a followed symlink
-	 * that redirects the walk into a module-excluded dir is refused. */
-	char abspath[MAXPATHLEN];
-};
-
-/* Append "/comp" to ds->abspath (no-op if it's unseeded/empty so non-daemon
- * callers pay nothing).  Returns -1 (ENAMETOOLONG) on overflow. */
-static int ds_path_push(struct dirstack *ds, const char *comp)
-{
-	size_t al = strlen(ds->abspath);
-	if (al == 0)
-		return 0;		/* unseeded: tracking disabled for this walk */
-	size_t cl = strlen(comp);
-	if (al + 1 + cl >= sizeof ds->abspath) {
-		errno = ENAMETOOLONG;
-		return -1;
-	}
-	ds->abspath[al] = '/';
-	memcpy(ds->abspath + al + 1, comp, cl + 1);
-	return 0;
-}
-
-/* Drop the last component of ds->abspath (mirrors a ".." pop). */
-static void ds_path_pop(struct dirstack *ds)
-{
-	char *slash;
-	if (!ds->abspath[0])
-		return;
-	slash = strrchr(ds->abspath, '/');
-	if (slash && slash != ds->abspath)
-		*slash = '\0';
-}
-
-/* Initialise with `anchor` (which may be AT_FDCWD) as the un-owned base.
- * Returns int for caller symmetry, but cannot fail (the fd array is inline). */
-static int ds_init(struct dirstack *ds, int anchor)
-{
-	ds->abspath[0] = '\0';
-	ds->fds[0] = anchor;
-	ds->top = 0;
-	return 0;
-}
-
-/* Close every pushed fd (but not the borrowed anchor at index 0). */
-static void ds_free(struct dirstack *ds)
-{
-	while (ds->top > 0)
-		close(ds->fds[ds->top--]);
-}
-
-static int ds_cur(struct dirstack *ds)
-{
-	return ds->fds[ds->top];
-}
-
-static int ds_push(struct dirstack *ds, int fd)
-{
-	if (ds->top + 1 >= DS_MAXDEPTH) {	/* deeper than we'll hold open */
-		close(fd);
-		errno = ENOMEM;
-		return -1;
-	}
-	ds->fds[++ds->top] = fd;
-	return 0;
-}
-
-/* Detach the current dir as an owned fd the caller must close.  At the anchor
- * (top 0) the anchor is borrowed, so return a fresh dup of it instead. */
-static int ds_take(struct dirstack *ds)
-{
-	if (ds->top > 0)
-		return ds->fds[ds->top--];
-	return openat(ds->fds[0], ".", O_RDONLY | O_DIRECTORY);
-}
-
-static int ds_walk_path(struct dirstack *ds, char *path, int *hops);
-
-/* Descend one path component on the stack: "." stays, ".." pops to the pinned
- * parent (ELOOP at the anchor), a real subdirectory is pushed, and an in-tree
- * directory symlink is followed by walking its (relative, possibly
- * ..-containing) target on the same stack.  Returns 0, or -1 with errno set:
- * ELOOP for a refused/escaping symlink or a hop overrun, otherwise the
- * underlying openat()/readlinkat() errno (ENOENT, a real ENOTDIR, EACCES). */
-static int ds_descend(struct dirstack *ds, const char *part, int *hops)
-{
-	if (part[0] == '.' && part[1] == '\0')
-		return 0;				/* "." -- no movement */
-	if (part[0] == '.' && part[1] == '.' && part[2] == '\0') {
-		if (ds->top == 0) {			/* would rise above the anchor */
-			errno = ELOOP;
-			return -1;
-		}
-		close(ds->fds[ds->top--]);		/* pop to the held parent fd */
-		ds_path_pop(ds);
-		return 0;
-	}
-
-	int fd = openat(ds_cur(ds), part, O_RDONLY | O_DIRECTORY | O_NOFOLLOW);
-	if (fd != -1) {					/* a real subdirectory */
-		if (ds_push(ds, fd) < 0)
-			return -1;
-		if (ds_path_push(ds, part) < 0)
-			return -1;
-		/* exclude-aware: refuse descending into a module-hidden dir (catches a
-		 * symlink that redirected the walk into an excluded subtree). */
-		if (abspath_outside_confinement(ds->abspath)) {
-			errno = ELOOP;
-			return -1;
-		}
-		return 0;
-	}
-	/* O_NOFOLLOW refused a symlink (NOFOLLOW_HIT_SYMLINK: ELOOP on Linux, EMLINK
-	 * on FreeBSD, EFTYPE on NetBSD/OpenBSD), or O_DIRECTORY hit a non-directory
-	 * (ENOTDIR).  Either may be a symlink, so fall through to the readlink probe;
-	 * anything else is a hard error. */
-	if (errno != ENOTDIR && !NOFOLLOW_HIT_SYMLINK(errno)) {
-		if (errno == EMFILE || errno == ENFILE) {
-			/* The resolver holds one dirfd per path component, so a deep path
-			 * can exhaust descriptors where plain open() would not.  Hint at
-			 * the fix once -- otherwise "Too many open files" is opaque. */
-			static int warned = 0;
-			if (!warned) {
-				int e = errno;
-				warned = 1;
-				rprintf(FWARNING, "out of file descriptors resolving a deep path;"
-					" raise the open-file limit (e.g. `ulimit -n`)\n");
-				errno = e;
-			}
-		}
-		return -1;
-	}
-	int open_errno = errno;
-
-	char buf[MAXPATHLEN];
-	ssize_t n = readlinkat(ds_cur(ds), part, buf, sizeof buf - 1);
-	if (n < 0) {
-		if (errno == EINVAL)			/* not a symlink: a real non-dir */
-			errno = open_errno;
-		return -1;
-	}
-	if (n == 0 || (size_t)n >= sizeof buf - 1) {
-		errno = ELOOP;				/* empty or truncated target */
-		return -1;
-	}
-	buf[n] = '\0';
-	if (buf[0] == '/') {				/* absolute target: refuse */
-		errno = ELOOP;
-		return -1;
-	}
-	if (--(*hops) < 0) {
-		errno = ELOOP;
-		return -1;
-	}
-	return ds_walk_path(ds, buf, hops);
-}
-
-/* Walk every component of a relative path on the stack (used for the basedir,
- * and for a followed symlink's target -- which may contain ".."). */
-static int ds_walk_path(struct dirstack *ds, char *path, int *hops)
-{
-	char *save = NULL;
-	for (char *c = strtok_r(path, "/", &save); c; c = strtok_r(NULL, "/", &save)) {
-		if (ds_descend(ds, c, hops) < 0)
-			return -1;
-	}
-	return 0;
-}
 
 /* Walk `relpath` confined beneath the borrowed anchor dirfd (which may be
  * AT_FDCWD) and return the opened leaf fd, or -1.  Does NOT close `anchor_fd` --
